Pit Manager.  For rF2 but capable of handling pit commands for R3E and iRacing (it would need a dispatch table for each and refactoring of the code to execute each command.  The voice input side is shared and all ready to go). The Pit event handler could also be called by Command Macro or the proposed Strategy Manager.
Uses autoFuelToEnd code lifted from iRacing code but then refactored.

A fair amount of testing carried out but any problems are likely to be in the event handler functions - there are a lot of corner cases (what can be changed by the pit menu varies from car to car) and controlling a UI designed for humans always throws up problems.

Very limited unit test but it does allow the new code to be exercised.